GHA-172: attempting to resolve artifact download
When workflow_run is designated in a workflow, the version that is
called is always from the default (main) branch, regardless of the
dependent workflow branch.  In other words, the dependent workflow might
trigger in the dev branch, but the triggered workflow with workflow_run
is always from the main branch.

[Ticket: https://newrelic.atlassian.net/browse/GHA-172]